+++ NEWS.md
@@ -1,5 +1,9 @@
 # infocausality 1.1
 
+### supersession noice
+
+* This package is now **superseded** by [`infoxtr`](https://CRAN.R-project.org/package=infoxtr).
+
 # infocausality 1.0
 
 ### new